# autogen/app/controllers.go
// Code generated by goagen v1.3.1, DO NOT EDIT.
//
// API "feedpushr": Application Controllers
//
// Command:
// $ goagen
// --design=github.com/ncarlier/feedpushr/design
// --out=$(GOPATH)/src/github.com/ncarlier/feedpushr/autogen
// --version=v1.3.1

package app

import (
	"context"
	"github.com/goadesign/goa"
	"github.com/goadesign/goa/cors"
	"net/http"
)

// initService sets up the service encoders, decoders and mux.
func initService(service *goa.Service) {
	// Setup encoders and decoders
	service.Encoder.Register(goa.NewJSONEncoder, "application/json")
	service.Encoder.Register(goa.NewGobEncoder, "application/gob", "application/x-gob")
	service.Encoder.Register(goa.NewXMLEncoder, "application/xml")
	service.Decoder.Register(goa.NewJSONDecoder, "application/json")
	service.Decoder.Register(goa.NewGobDecoder, "application/gob", "application/x-gob")
	service.Decoder.Register(goa.NewXMLDecoder, "application/xml")

	// Setup default encoder and decoder
	service.Encoder.Register(goa.NewJSONEncoder, "*/*")
	service.Decoder.Register(goa.NewJSONDecoder, "*/*")
}

// FeedController is the controller interface for the Feed actions.
type FeedController interface {
	goa.Muxer
	Create(*CreateFeedContext) error
	Delete(*DeleteFeedContext) error
	Get(*GetFeedContext) error
	List(*ListFeedContext) error
	Start(*StartFeedContext) error
	Stop(*StopFeedContext) error
}

// MountFeedController "mounts" a Feed resource controller on the given service.
func MountFeedController(service *goa.Service, ctrl FeedController) {
	initService(service)
	var h goa.Handler
	service.Mux.Handle("OPTIONS", "/v1/feeds", ctrl.MuxHandler("preflight", handleFeedOrigin(cors.HandlePreflight()), nil))
	service.Mux.Handle("OPTIONS", "/v1/feeds/:id", ctrl.MuxHandler("preflight", handleFeedOrigin(cors.HandlePreflight()), nil))
	service.Mux.Handle("OPTIONS", "/v1/feeds/:id/start", ctrl.MuxHandler("preflight", handleFeedOrigin(cors.HandlePreflight()), nil))
	service.Mux.Handle("OPTIONS", "/v1/feeds/:id/stop", ctrl.MuxHandler("preflight", handleFeedOrigin(cors.HandlePreflight()), nil))

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewCreateFeedContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Create(rctx)
	}
	h = handleFeedOrigin(h)
	service.Mux.Handle("POST", "/v1/feeds", ctrl.MuxHandler("create", h, nil))
	service.LogInfo("mount", "ctrl", "Feed", "action", "Create", "route", "POST /v1/feeds")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewDeleteFeedContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Delete(rctx)
	}
	h = handleFeedOrigin(h)
	service.Mux.Handle("DELETE", "/v1/feeds/:id", ctrl.MuxHandler("delete", h, nil))
	service.LogInfo("mount", "ctrl", "Feed", "action", "Delete", "route", "DELETE /v1/feeds/:id")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewGetFeedContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Get(rctx)
	}
	h = handleFeedOrigin(h)
	service.Mux.Handle("GET", "/v1/feeds/:id", ctrl.MuxHandler("get", h, nil))
	service.LogInfo("mount", "ctrl", "Feed", "action", "Get", "route", "GET /v1/feeds/:id")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewListFeedContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.List(rctx)
	}
	h = handleFeedOrigin(h)
	service.Mux.Handle("GET", "/v1/feeds", ctrl.MuxHandler("list", h, nil))
	service.LogInfo("mount", "ctrl", "Feed", "action", "List", "route", "GET /v1/feeds")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewStartFeedContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Start(rctx)
	}
	h = handleFeedOrigin(h)
	service.Mux.Handle("POST", "/v1/feeds/:id/start", ctrl.MuxHandler("start", h, nil))
	service.LogInfo("mount", "ctrl", "Feed", "action", "Start", "route", "POST /v1/feeds/:id/start")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewStopFeedContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Stop(rctx)
	}
	h = handleFeedOrigin(h)
	service.Mux.Handle("POST", "/v1/feeds/:id/stop", ctrl.MuxHandler("stop", h, nil))
	service.LogInfo("mount", "ctrl", "Feed", "action", "Stop", "route", "POST /v1/feeds/:id/stop")
}

// handleFeedOrigin applies the CORS response headers corresponding to the origin.
func handleFeedOrigin(h goa.Handler) goa.Handler {

	return func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		origin := req.Header.Get("Origin")
		if origin == "" {
			// Not a CORS request
			return h(ctx, rw, req)
		}
		if cors.MatchOrigin(origin, "http://swagger.goa.design") {
			ctx = goa.WithLogContext(ctx, "origin", origin)
			rw.Header().Set("Access-Control-Allow-Origin", origin)
			rw.Header().Set("Vary", "Origin")
			rw.Header().Set("Access-Control-Max-Age", "600")
			rw.Header().Set("Access-Control-Allow-Credentials", "true")
			if acrm := req.Header.Get("Access-Control-Request-Method"); acrm != "" {
				// We are handling a preflight request
				rw.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT, PATCH, DELETE")
			}
			return h(ctx, rw, req)
		}

		return h(ctx, rw, req)
	}
}

// HealthController is the controller interface for the Health actions.
type HealthController interface {
	goa.Muxer
	Get(*GetHealthContext) error
}

// MountHealthController "mounts" a Health resource controller on the given service.
func MountHealthController(service *goa.Service, ctrl HealthController) {
	initService(service)
	var h goa.Handler
	service.Mux.Handle("OPTIONS", "/v1/healthz", ctrl.MuxHandler("preflight", handleHealthOrigin(cors.HandlePreflight()), nil))

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewGetHealthContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Get(rctx)
	}
	h = handleHealthOrigin(h)
	service.Mux.Handle("GET", "/v1/healthz", ctrl.MuxHandler("get", h, nil))
	service.LogInfo("mount", "ctrl", "Health", "action", "Get", "route", "GET /v1/healthz")
}

// handleHealthOrigin applies the CORS response headers corresponding to the origin.
func handleHealthOrigin(h goa.Handler) goa.Handler {

	return func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		origin := req.Header.Get("Origin")
		if origin == "" {
			// Not a CORS request
			return h(ctx, rw, req)
		}
		if cors.MatchOrigin(origin, "http://swagger.goa.design") {
			ctx = goa.WithLogContext(ctx, "origin", origin)
			rw.Header().Set("Access-Control-Allow-Origin", origin)
			rw.Header().Set("Vary", "Origin")
			rw.Header().Set("Access-Control-Max-Age", "600")
			rw.Header().Set("Access-Control-Allow-Credentials", "true")
			if acrm := req.Header.Get("Access-Control-Request-Method"); acrm != "" {
				// We are handling a preflight request
				rw.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT, PATCH, DELETE")
			}
			return h(ctx, rw, req)
		}

		return h(ctx, rw, req)
	}
}

// OpmlController is the controller interface for the Opml actions.
type OpmlController interface {
	goa.Muxer
	Get(*GetOpmlContext) error
	Upload(*UploadOpmlContext) error
}

// MountOpmlController "mounts" a Opml resource controller on the given service.
func MountOpmlController(service *goa.Service, ctrl OpmlController) {
	initService(service)
	var h goa.Handler
	service.Mux.Handle("OPTIONS", "/v1/opml", ctrl.MuxHandler("preflight", handleOpmlOrigin(cors.HandlePreflight()), nil))

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewGetOpmlContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Get(rctx)
	}
	h = handleOpmlOrigin(h)
	service.Mux.Handle("GET", "/v1/opml", ctrl.MuxHandler("get", h, nil))
	service.LogInfo("mount", "ctrl", "Opml", "action", "Get", "route", "GET /v1/opml")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewUploadOpmlContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Upload(rctx)
	}
	h = handleOpmlOrigin(h)
	service.Mux.Handle("POST", "/v1/opml", ctrl.MuxHandler("upload", h, nil))
	service.LogInfo("mount", "ctrl", "Opml", "action", "Upload", "route", "POST /v1/opml")
}

// handleOpmlOrigin applies the CORS response headers corresponding to the origin.
func handleOpmlOrigin(h goa.Handler) goa.Handler {

	return func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		origin := req.Header.Get("Origin")
		if origin == "" {
			// Not a CORS request
			return h(ctx, rw, req)
		}
		if cors.MatchOrigin(origin, "http://swagger.goa.design") {
			ctx = goa.WithLogContext(ctx, "origin", origin)
			rw.Header().Set("Access-Control-Allow-Origin", origin)
			rw.Header().Set("Vary", "Origin")
			rw.Header().Set("Access-Control-Max-Age", "600")
			rw.Header().Set("Access-Control-Allow-Credentials", "true")
			if acrm := req.Header.Get("Access-Control-Request-Method"); acrm != "" {
				// We are handling a preflight request
				rw.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT, PATCH, DELETE")
			}
			return h(ctx, rw, req)
		}

		return h(ctx, rw, req)
	}
}

// PshbController is the controller interface for the Pshb actions.
type PshbController interface {
	goa.Muxer
	Pub(*PubPshbContext) error
	Sub(*SubPshbContext) error
}

// MountPshbController "mounts" a Pshb resource controller on the given service.
func MountPshbController(service *goa.Service, ctrl PshbController) {
	initService(service)
	var h goa.Handler
	service.Mux.Handle("OPTIONS", "/v1/pshb", ctrl.MuxHandler("preflight", handlePshbOrigin(cors.HandlePreflight()), nil))

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewPubPshbContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Pub(rctx)
	}
	h = handlePshbOrigin(h)
	service.Mux.Handle("POST", "/v1/pshb", ctrl.MuxHandler("pub", h, nil))
	service.LogInfo("mount", "ctrl", "Pshb", "action", "Pub", "route", "POST /v1/pshb")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewSubPshbContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Sub(rctx)
	}
	h = handlePshbOrigin(h)
	service.Mux.Handle("GET", "/v1/pshb", ctrl.MuxHandler("sub", h, nil))
	service.LogInfo("mount", "ctrl", "Pshb", "action", "Sub", "route", "GET /v1/pshb")
}

// handlePshbOrigin applies the CORS response headers corresponding to the origin.
func handlePshbOrigin(h goa.Handler) goa.Handler {

	return func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		origin := req.Header.Get("Origin")
		if origin == "" {
			// Not a CORS request
			return h(ctx, rw, req)
		}
		if cors.MatchOrigin(origin, "http://swagger.goa.design") {
			ctx = goa.WithLogContext(ctx, "origin", origin)
			rw.Header().Set("Access-Control-Allow-Origin", origin)
			rw.Header().Set("Vary", "Origin")
			rw.Header().Set("Access-Control-Max-Age", "600")
			rw.Header().Set("Access-Control-Allow-Credentials", "true")
			if acrm := req.Header.Get("Access-Control-Request-Method"); acrm != "" {
				// We are handling a preflight request
				rw.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT, PATCH, DELETE")
			}
			return h(ctx, rw, req)
		}

		return h(ctx, rw, req)
	}
}

// SwaggerController is the controller interface for the Swagger actions.
type SwaggerController interface {
	goa.Muxer
	goa.FileServer
}

// MountSwaggerController "mounts" a Swagger resource controller on the given service.
func MountSwaggerController(service *goa.Service, ctrl SwaggerController) {
	initService(service)
	var h goa.Handler
	service.Mux.Handle("OPTIONS", "/swagger.json", ctrl.MuxHandler("preflight", handleSwaggerOrigin(cors.HandlePreflight()), nil))

	h = ctrl.FileHandler("/swagger.json", "var/public/swagger.json")
	h = handleSwaggerOrigin(h)
	service.Mux.Handle("GET", "/swagger.json", ctrl.MuxHandler("serve", h, nil))
	service.LogInfo("mount", "ctrl", "Swagger", "files", "var/public/swagger.json", "route", "GET /swagger.json")
}

// handleSwaggerOrigin applies the CORS response headers corresponding to the origin.
func handleSwaggerOrigin(h goa.Handler) goa.Handler {

	return func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		origin := req.Header.Get("Origin")
		if origin == "" {
			// Not a CORS request
			return h(ctx, rw, req)
		}
		if cors.MatchOrigin(origin, "*") {
			ctx = goa.WithLogContext(ctx, "origin", origin)
			rw.Header().Set("Access-Control-Allow-Origin", origin)
			rw.Header().Set("Access-Control-Allow-Credentials", "false")
			if acrm := req.Header.Get("Access-Control-Request-Method"); acrm != "" {
				// We are handling a preflight request
				rw.Header().Set("Access-Control-Allow-Methods", "GET, OPTIONS")
			}
			return h(ctx, rw, req)
		}
		if cors.MatchOrigin(origin, "http://swagger.goa.design") {
			ctx = goa.WithLogContext(ctx, "origin", origin)
			rw.Header().Set("Access-Control-Allow-Origin", origin)
			rw.Header().Set("Vary", "Origin")
			rw.Header().Set("Access-Control-Max-Age", "600")
			rw.Header().Set("Access-Control-Allow-Credentials", "true")
			if acrm := req.Header.Get("Access-Control-Request-Method"); acrm != "" {
				// We are handling a preflight request
				rw.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT, PATCH, DELETE")
			}
			return h(ctx, rw, req)
		}

		return h(ctx, rw, req)
	}
}

// VarsController is the controller interface for the Vars actions.
type VarsController interface {
	goa.Muxer
	Get(*GetVarsContext) error
}

// MountVarsController "mounts" a Vars resource controller on the given service.
func MountVarsController(service *goa.Service, ctrl VarsController) {
	initService(service)
	var h goa.Handler
	service.Mux.Handle("OPTIONS", "/v1/vars", ctrl.MuxHandler("preflight", handleVarsOrigin(cors.HandlePreflight()), nil))

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewGetVarsContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Get(rctx)
	}
	h = handleVarsOrigin(h)
	service.Mux.Handle("GET", "/v1/vars", ctrl.MuxHandler("get", h, nil))
	service.LogInfo("mount", "ctrl", "Vars", "action", "Get", "route", "GET /v1/vars")
}

// handleVarsOrigin applies the CORS response headers corresponding to the origin.
func handleVarsOrigin(h goa.Handler) goa.Handler {

	return func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		origin := req.Header.Get("Origin")
		if origin == "" {
			// Not a CORS request
			return h(ctx, rw, req)
		}
		if cors.MatchOrigin(origin, "http://swagger.goa.design") {
			ctx = goa.WithLogContext(ctx, "origin", origin)
			rw.Header().Set("Access-Control-Allow-Origin", origin)
			rw.Header().Set("Vary", "Origin")
			rw.Header().Set("Access-Control-Max-Age", "600")
			rw.Header().Set("Access-Control-Allow-Credentials", "true")
			if acrm := req.Header.Get("Access-Control-Request-Method"); acrm != "" {
				// We are handling a preflight request
				rw.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT, PATCH, DELETE")
			}
			return h(ctx, rw, req)
		}

		return h(ctx, rw, req)
	}
}