# var/public/swagger.json
{"swagger":"2.0","info":{"title":"The feedpushr API","description":"A simple feed aggregator daemon with sugar on top.","contact":{"name":"Nicolas Carlier","url":"https://github.com/ncarlier"},"license":{"name":"MIT","url":"https://github.com/goadesign/goa/blob/master/LICENSE"},"version":""},"host":"localhost:8080","schemes":["http"],"consumes":["application/json","application/xml","application/gob","application/x-gob"],"produces":["application/json","application/xml","application/gob","application/x-gob"],"paths":{"/swagger.json":{"get":{"summary":"Download var/public/swagger.json","operationId":"swagger#/swagger.json","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}},"schemes":["http"]}},"/v1/feeds":{"get":{"tags":["feed"],"summary":"list feed","description":"Retrieve all feeds","operationId":"feed#list","produces":["application/vnd.feedpushr.feed.v1+json; type=collection","application/vnd.goa.error"],"parameters":[{"name":"limit","in":"query","description":"Fetch limit","required":false,"type":"integer","default":10,"maximum":100,"minimum":1},{"name":"page","in":"query","description":"Page to fetch","required":false,"type":"integer","default":1,"minimum":1}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/FeedCollection"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"404":{"description":"Not Found"}},"schemes":["http"]},"post":{"tags":["feed"],"summary":"create feed","description":"Create a new feed","operationId":"feed#create","produces":["application/vnd.goa.error"],"parameters":[{"name":"url","in":"query","description":"Feed URL","required":true,"type":"string","format":"uri"}],"responses":{"201":{"description":"Resource created","headers":{"Location":{"description":"href to created resource","type":"string","pattern":"/feeds/[0-9a-f]+"}}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}},"/v1/feeds/{id}":{"get":{"tags":["feed"],"summary":"get feed","description":"Retrieve feed with given id","operationId":"feed#get","produces":["application/vnd.feedpushr.feed.v1+json","application/vnd.goa.error"],"parameters":[{"name":"id","in":"path","description":"Feed ID","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Feed"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"404":{"description":"Not Found"}},"schemes":["http"]},"delete":{"tags":["feed"],"summary":"delete feed","description":"Delete a feed","operationId":"feed#delete","produces":["application/vnd.goa.error"],"parameters":[{"name":"id","in":"path","description":"Feed ID","required":true,"type":"string"}],"responses":{"204":{"description":"No Content"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"404":{"description":"Not Found"}},"schemes":["http"]}},"/v1/feeds/{id}/start":{"post":{"tags":["feed"],"summary":"start feed","description":"Start feed aggregation","operationId":"feed#start","produces":["application/vnd.goa.error"],"parameters":[{"name":"id","in":"path","required":true,"type":"string"}],"responses":{"202":{"description":"Accepted"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"404":{"description":"Not Found"}},"schemes":["http"]}},"/v1/feeds/{id}/stop":{"post":{"tags":["feed"],"summary":"stop feed","description":"Stop feed aggregation","operationId":"feed#stop","produces":["application/vnd.goa.error"],"parameters":[{"name":"id","in":"path","required":true,"type":"string"}],"responses":{"202":{"description":"Accepted"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"404":{"description":"Not Found"}},"schemes":["http"]}},"/v1/healthz":{"get":{"tags":["health"],"summary":"get health","description":"Perform health check.","operationId":"health#get","produces":["text/plain"],"responses":{"200":{"description":"OK"}},"schemes":["http"]}},"/v1/opml":{"get":{"tags":["opml"],"summary":"get opml","description":"Get all feeds as an OMPL format","operationId":"opml#get","produces":["application/vnd.goa.error","text/plain"],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]},"post":{"tags":["opml"],"summary":"upload opml","description":"Upload an OPML file to create feeds","operationId":"opml#upload","produces":["application/vnd.goa.error"],"responses":{"201":{"description":"Resource created","headers":{"Location":{"description":"href to created resource","type":"string","pattern":"/feeds"}}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}},"/v1/pshb":{"get":{"tags":["pshb"],"summary":"sub pshb","description":"Callback to validate the (un)subscription to the topic of a Hub","operationId":"pshb#sub","produces":["application/vnd.goa.error","text/plain"],"parameters":[{"name":"hub.challenge","in":"query","description":"A hub-generated random string","required":true,"type":"string"},{"name":"hub.lease_seconds","in":"query","description":"The hub-determined number of seconds that the subscription will stay active before expiring","required":false,"type":"integer"},{"name":"hub.mode","in":"query","description":"The literal string \"subscribe\" or \"unsubscribe\"","required":true,"type":"string","enum":["subscribe","unsubscribe"]},{"name":"hub.topic","in":"query","description":"The topic URL given in the corresponding subscription request","required":true,"type":"string","format":"uri"}],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]},"post":{"tags":["pshb"],"summary":"pub pshb","description":"Publication endpoint for PSHB hubs","operationId":"pshb#pub","produces":["application/vnd.goa.error","text/plain"],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}},"/v1/vars":{"get":{"tags":["vars"],"summary":"get vars","description":"Get all internals exp vars","operationId":"vars#get","produces":["application/vnd.goa.error","text/plain"],"responses":{"200":{"description":"OK"},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}}},"definitions":{"Feed":{"title":"Mediatype identifier: application/vnd.feedpushr.feed.v1+json; view=default","type":"object","properties":{"cdate":{"type":"string","description":"Date of creation","example":"2010-10-23T13:34:07Z","format":"date-time"},"errorCount":{"type":"integer","description":"Number of consecutive aggregation errors","example":333691758294841896,"format":"int64"},"errorMsg":{"type":"string","description":"Last aggregation error","example":"Est suscipit."},"htmlUrl":{"type":"string","description":"URL of the feed website","example":"http://www.hashicorp.com/blog"},"hubUrl":{"type":"string","description":"URL of the PubSubHubbud hub","example":"http://pubsubhubbub.appspot.com"},"id":{"type":"string","description":"ID of feed (MD5 of the xmlUrl)","example":"5bfb841c028281c0051828c115fd1f50"},"lastCheck":{"type":"string","description":"Last aggregation pass","example":"1993-05-15T16:34:02Z","format":"date-time"},"mdate":{"type":"string","description":"Date of modification","example":"1981-09-22T23:39:59Z","format":"date-time"},"nextCheck":{"type":"string","description":"Next aggregation pass","example":"1975-05-22T16:26:55Z","format":"date-time"},"text":{"type":"string","description":"Text attribute of the Feed","example":"RSS Feed"},"title":{"type":"string","description":"Title of the Feed","example":"Hashicorp Blog"},"xmlUrl":{"type":"string","description":"URL of the XML feed","example":"http://www.hashicorp.com/feed.xml"}},"description":"A RSS feed (default view)","example":{"cdate":"2010-10-23T13:34:07Z","errorCount":333691758294841896,"errorMsg":"Est suscipit.","htmlUrl":"http://www.hashicorp.com/blog","hubUrl":"http://pubsubhubbub.appspot.com","id":"5bfb841c028281c0051828c115fd1f50","lastCheck":"1993-05-15T16:34:02Z","mdate":"1981-09-22T23:39:59Z","nextCheck":"1975-05-22T16:26:55Z","text":"RSS Feed","title":"Hashicorp Blog","xmlUrl":"http://www.hashicorp.com/feed.xml"},"required":["id","xmlUrl","title","cdate","mdate"]},"FeedCollection":{"title":"Mediatype identifier: application/vnd.feedpushr.feed.v1+json; type=collection; view=default","type":"array","items":{"$ref":"#/definitions/Feed"},"description":"FeedCollection is the media type for an array of Feed (default view)","example":[{"cdate":"2010-10-23T13:34:07Z","errorCount":333691758294841896,"errorMsg":"Est suscipit.","htmlUrl":"http://www.hashicorp.com/blog","hubUrl":"http://pubsubhubbub.appspot.com","id":"5bfb841c028281c0051828c115fd1f50","lastCheck":"1993-05-15T16:34:02Z","mdate":"1981-09-22T23:39:59Z","nextCheck":"1975-05-22T16:26:55Z","text":"RSS Feed","title":"Hashicorp Blog","xmlUrl":"http://www.hashicorp.com/feed.xml"},{"cdate":"2010-10-23T13:34:07Z","errorCount":333691758294841896,"errorMsg":"Est suscipit.","htmlUrl":"http://www.hashicorp.com/blog","hubUrl":"http://pubsubhubbub.appspot.com","id":"5bfb841c028281c0051828c115fd1f50","lastCheck":"1993-05-15T16:34:02Z","mdate":"1981-09-22T23:39:59Z","nextCheck":"1975-05-22T16:26:55Z","text":"RSS Feed","title":"Hashicorp Blog","xmlUrl":"http://www.hashicorp.com/feed.xml"}]},"error":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"code":{"type":"string","description":"an application-specific error code, expressed as a string value.","example":"invalid_value"},"detail":{"type":"string","description":"a human-readable explanation specific to this occurrence of the problem.","example":"Value of ID must be an integer"},"id":{"type":"string","description":"a unique identifier for this particular occurrence of the problem.","example":"3F1FKVRR"},"meta":{"type":"object","description":"a meta object containing non-standard meta-information about the error.","example":{"timestamp":1458609066},"additionalProperties":true},"status":{"type":"string","description":"the HTTP status code applicable to this problem, expressed as a string value.","example":"400"}},"description":"Error response media type (default view)","example":{"code":"invalid_value","detail":"Value of ID must be an integer","id":"3F1FKVRR","meta":{"timestamp":1458609066},"status":"400"}}},"responses":{"Accepted":{"description":"Accepted"},"NoContent":{"description":"No Content"},"NotFound":{"description":"Not Found"},"OK":{"description":"OK","schema":{"$ref":"#/definitions/Feed"}}},"externalDocs":{"description":"feedpushr guide","url":"https://github.com/ncarlier/feedpusher/README.md"}}